+++ src/Tecan/ReagentDistribution/ReagentDistributionDirection.php
@@ -7,16 +7,11 @@ use BenSampo\Enum\Enum;
 /**
  * @method static static LEFT_TO_RIGHT()
  * @method static static RIGHT_TO_LEFT()
+ *
+ * @extends Enum<int>
  */
 final class ReagentDistributionDirection extends Enum
 {
-    public const LEFT_TO_RIGHT = '0';
-    public const RIGHT_TO_LEFT = '1';
-
-    public function toString(): string
-    {
-        assert(is_string($this->value));
-
-        return $this->value;
-    }
+    public const LEFT_TO_RIGHT = 0;
+    public const RIGHT_TO_LEFT = 1;
 }